feat: add new file for comp Timer
